+++ packages/tsurlfilter/CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## Unreleased
 
 ### Added
+- New `$to` modifier to match requests by target domains and subdomains.
 - In the MV3 declarative converter, the obsolete `$mp4` and `$empty` are now
   converted by replacing them with `$redirect` rules.